--- a/backend/jest-mongodb-config.js
+++ b/backend/jest-mongodb-config.js
@@ -1,7 +1,7 @@
 module.exports = {
   mongodbMemoryServerOptions: {
     binary: {
-      version: "4.0.3",
+      version: "6.0.12",
       skipMD5: true,
     },
     instance: {