--- a/backend/src/dal/leaderboards.ts
+++ b/backend/src/dal/leaderboards.ts
@@ -192,12 +192,6 @@ export async function update(
         },
       },
       {
-        $addFields: {
-          type: "speedStats",
-        },
-      },
-
-      {
         $merge: {
           into: "public",
           on: "_id",