--- a/backend/src/dal/public.ts
+++ b/backend/src/dal/public.ts
@@ -7,7 +7,7 @@ export async function updateStats(
   time: number
 ): Promise<boolean> {
   await db.collection<MonkeyTypes.PublicStats>("public").updateOne(
-    { type: "stats" },
+    { _id: "stats" },
     {
       $inc: {
         testsCompleted: 1,
@@ -40,7 +40,7 @@ export async function getSpeedHistogram(
 export async function getTypingStats(): Promise<MonkeyTypes.PublicStats> {
   const stats = await db
     .collection<MonkeyTypes.PublicStats>("public")
-    .findOne({ type: "stats" }, { projection: { _id: 0 } });
+    .findOne({ _id: "stats" }, { projection: { _id: 0 } });
   if (!stats) {
     throw new MonkeyError(
       404,